auth/clientcredentials: add HTTP middleware
